Move errors utils
