+++ universal-login-commons/lib/core/utils/errors/heplers.ts
