+++ universal-login-commons/lib/index.ts
@@ -24,7 +24,7 @@ export {getEmojiColor, getEmojiNumber, CATEGORIES, getBaseEmojiCode, getEmojiSet
 export {generateBackupCode} from './core/utils/generateBackupCode';
 export {getEmojiCodePoint} from './core/utils/emojiCodePoint';
 export {safeMultiply} from './core/utils/safeMultiply';
-export {ensure, ensureNotNull, ensureNotEmpty, onCritical} from './core/utils/errors';
+export {ensure, ensureNotNull, ensureNotEmpty, onCritical} from './core/utils/errors/heplers';
 export {computeCounterfactualAddress, computeContractAddress} from './core/utils/contracts/computeContractAddress';
 export {BalanceChecker} from './integration/ethereum/BalanceChecker';
 export {RequiredBalanceChecker} from './core/services/RequiredBalanceChecker';
